backend/requirements.txt
flask
flask-restx
flask_cors
Flask-PyMongo

fastapi
uvicorn[standard]
motor
pymongo
feedparser
python-dotenv